requirements.txt
adtk==0.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Faker==15.1.1
fonttools==4.37.4
idna==3.4
influxdb==5.3.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
msgpack==1.0.4
numpy==1.23.3
packaging==21.3
paho-mqtt==1.6.1
pandas==1.5.1
patsy==0.5.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
statsmodels==0.13.2
tabulate==0.9.0
threadpoolctl==3.1.0
urllib3==1.26.12
